fix message receiver

+++ EmotionDetector/Shared/EmotionResult.cs
@@ -12,7 +12,7 @@ namespace Shared
         public Guid SessionId { get; set; }
         public DateTime Date { get; set; }
         public string Emotion { get; set; }
-        public double Score { get; set; }
+        public int Score { get; set; }
 
     }
 }
